test(ci): split tool benchmark workflow assertions into their own file
